+++ packages/rs-dpp/src/tests/version/protocol_version_validator_spec.rs
@@ -2,7 +2,7 @@ use std::collections::HashMap;
 use crate::errors::consensus::{ConsensusError};
 use crate::version::{ProtocolVersionValidator};
 
-pub fn setup_test() -> (u32, u32, HashMap<u32, u32>) {
+pub fn setup_test() -> (u64, u64, HashMap<u64, u64>) {
     let current_protocol_version = 1;
     let latest_protocol_version = 1;
     let mut version_compatibility_map = HashMap::new();
